revert fix of suspicious int casting in getTotalxp()

+++ src/main/java/net/lyof/sortilege/utils/XPHelper.java
@@ -48,7 +48,7 @@ public class XPHelper {
 
         for (int i = 0; i <= level; i++) {
             dummy.experienceLevel = i;
-            total += (int) (dummy.getXpNeededForNextLevel() * (i == level ? progress : 1));
+            total += dummy.getXpNeededForNextLevel() * (i == level ? progress : 1);
         }
 
         totalxpCache.put(level, total);
